call api update user info
