+++ package.json
@@ -33,6 +33,7 @@
         "react-router-dom": "^6.0.2",
         "react-scripts": "5.0.0",
         "react-syntax-highlighter": "15.4.5",
+        "react-toastify": "^8.2.0",
         "tw-elements": "^1.0.0-alpha12",
         "web-vitals": "1.0.1"
     },
