+++ src/routes.js
@@ -8,7 +8,6 @@ import ProductListPages from "pages/LandingPages/ProductListPage";
 import SignUp from "pages/LandingPages/SignUp";
 import ShoppingCart from "pages/LandingPages/ShoppingCart";
 import UserProfile from "pages/LandingPages/UserProfile";
-import PrivateRoute from "./pages/PrivateRoute";
 
 const routes = [
   {
@@ -38,9 +37,7 @@ const routes = [
           {
             name: "Profile",
             route: "/user-profile",
-            component: <PrivateRoute>
-                          <UserProfile />
-                      </PrivateRoute>,
+            component: <UserProfile />,
           },
         ],
       },
